Change ODataPackage so its Published property is included in output
In commit 9a2dd68 an EntityPropertyMapping attribute was added to the ODataPackage class which removed the published date from the OData output, causing Visual Studio to not display a published date for packages.  This changes the EntityPropertyMapping attribute to include the published date in the OData output again so Visual Studio can display a published date for packages.